requirements.txt
# Core Dependencies
PySide6>=6.6.0                    # GUI framework
sentence-transformers>=2.2.2      # Embeddings (BGE-small)
chromadb>=0.4.22                  # Vector database
watchdog>=3.0.0                   # File system monitoring
numpy<2.0                         # Pin to 1.x for compatibility

# Document Processing
pypdf>=3.17.0                     # PDF reading
python-docx>=1.1.0                # DOCX reading
pytesseract>=0.3.10               # OCR
Pillow>=10.1.0                    # Image processing

# Optional: Local LLM Support
llama-cpp-python>=0.2.27          # Local GGUF models (optional)

# Optional: API LLM Support
openai>=1.6.1                     # OpenAI API (optional)
anthropic>=0.8.1                  # Anthropic API (optional)

# Utilities
numpy>=1.24.0                     # Numerical operations
tqdm>=4.66.0                      # Progress bars

# Note: For GPU acceleration (optional):
# - sentence-transformers with GPU: Install PyTorch with CUDA
# - llama-cpp-python with GPU: pip install llama-cpp-python[cuda]
#
# For CPU-only installation, the above requirements are sufficient.